refactor: move the definition of `StackFrame` to `stack_frame.rs`
